# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Movies Battle
  description: |-
    This is a sample card game API Server. The main rule of our game consists in to provide the user two options of differente movie titles and they must guess which one has the best rating on [IMDb](https://www.omdbapi.com/).
    
  contact:
    email: italo.costa.ce@gmail.com
  version: 1.0.0
tags:
  - name: Game
    description: Managing the game session
  - name: Quiz
    description: Check and answer your next quiz
paths:
  /games:
    post:
      tags:
        - Game
      summary: Start a new game session
      description: Start a new game session to see your next quiz
      operationId: startGame
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: You already have a session running up. Please, stop the current one or continue playing with that.
        '403':
          description: User not found
      security:
        - moviesbattle_auth: []
    delete:
      tags:
        - Game
      summary: Stop the current game session
      description: Stop the current game session, if you have one, and see the fina results.
      operationId: stopGame
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: You don't have any session running up. Please, start a new one.
        '403':
          description: User not found
      security:
        - moviesbattle_auth: []
    get:
      tags:
        - Game
      summary: Find your game session information
      description: Find your game session information, if you have one.
      operationId: getGame
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: You don't have any session running up. Please, start a new one.
        '403':
          description: User not found
      security:
        - moviesbattle_auth: []
  /games/ranking:
    get:
      tags:
        - Game
      summary: See the game ranking
      operationId: getGameRanking
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ranking'
        '403':
          description: User not found
      security:
        - moviesbattle_auth: []
  /quizzes:
    get:
      tags:
        - Quiz
      summary: Check your next quiz information
      operationId: getQuiz
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quiz'
        '400':
          description: You don't have any session running up. Please, start a new one.
        '403':
          description: User not found
      security:
        - moviesbattle_auth: []
    post:
      tags:
        - Quiz
      summary: Answer your quiz
      operationId: replyQuiz
      parameters:
        - name: selectedOption
          in: query
          description: The selected quiz option
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
                example: Incorrect answer!
        '400':
          description: You don't have any quiz waiting for your reply. Please, start a new one.
        '403':
          description: User not found
      security:
        - moviesbattle_auth: []
components:
  schemas:
    Ranking:
      type: object
      properties:
        user:
          type: string
          example: italo.costa
        score:
          type: number
          format: float
          example: 4.0
    Game:
      type: object
      properties:
        user:
          type: string
          example: italo.costa
        score:
          type: number
          format: float
          example: 5.4
        amountCorrectQuiz:
          type: integer
          format: int32
          example: 12
        amountWrongQuiz:
          type: integer
          format: int32
          example: 2
        remainingFailures:
          type: integer
          format: int32
          example: 3
        creationTime:
          type: string
          format: date-time
        lastUpdateTime:
          type: string
          format: date-time
    Quiz:
      type: object
      properties:
        option:
          type: integer
          format: int32
          example: 2
        title:
          type: string
          example: The Godfather
  securitySchemes:
    moviesbattle_auth:
      type: http
      scheme: basic